requirements.txt
google-generativeai
python-dotenv
boto3
requests
pillow
pandas
PyPDF2
streamlit
streamlit-authenticator